backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
torch
deep-translator
gtts
pydub
jinja2
aiofiles
audioop-lts
requests
pydantic-settings
pillow
pytesseract
reportlab
python-docx
opencv-python-headless
rapidocr-onnxruntime
